+++ R/rbacon-package.R
@@ -12,8 +12,8 @@
 #' @importFrom Rcpp evalCpp
 #' @importFrom coda as.mcmc gelman.diag mcmc.list
 #' @importFrom data.table fread fwrite
-#' @importFrom rice draw.dates
-#' @importFrom rintcal ccurve
+#' @importFrom rice draw.dates F14CtoC14 pMCtoC14
+#' @importFrom rintcal ccurve mix.curves new.ccdir
 #' @useDynLib rbacon, .registration=TRUE
 #' @name rbacon
 #' @keywords internal