+++ R/read_write.R
@@ -354,6 +354,10 @@ read.dets <- function(core, coredir, othername=c(), set=get('info'), sep=",", de
     dets[,1] <- sub(",", "", dets[,1])
     message("Warning, I removed commas from the lab ID fields")
   }
+  if(length(grep('"', dets)) > 0)  {# commas in the lab IDs, which can confuse bacon (c++)
+    dets[,1] <- sub('"', '', dets[,1])
+    message("Warning, I removed quotation marks")
+  }
 
   # if current dets differ from original .csv file, rewrite it
   if(changed > 0)