+++ NEWS.md
@@ -1,3 +1,6 @@
+# rbacon 3.3.2
+* radiocarbon ages can now be entered as F14C or pMC values (e.g. for postbomb dates). Please indicate using the F14C or pMC option which ones are in F14C/pMC (e.g., Bacon(F14C=2:4) if dates 2 to 4 are in F14C). These values will then automatically be rewritten as C14 ages.
+
 # rbacon 3.3.1
 * now uses the updated rice package (which loads the data rintcal package).
 * new options in the age-depth function to steer the formatting of the prior information, such as line width and colour